♻️ refactor: NewPost, EditPost에서 사용할 공통 입력양식을 컴포넌트로 분리
